Completed the GenericObjectField method
Object picker filters types propertly and the type name is shown correctly. Apparently, object picker works just like AssetDatabase filter and filter by type name, so a unique name for each concrete class was required. Assembly GUID doesn't change if the type is updated, so it was the best option for naming the concrete class